# backend/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "sCalenderPlus API",
    "description": "Public REST API of sCalenderPlus. See docs/architecture/api.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/me/two-factor": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Two-factor authentication status",
        "operationId": "GetTwoFactorStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatusResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/me/two-factor/setup": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Create a new authenticator secret (QR code URI); 409 while 2FA is enabled",
        "operationId": "SetUpTwoFactor",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/me/two-factor/enable": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Turn on 2FA with a current authenticator code; returns the recovery codes once",
        "operationId": "EnableTwoFactor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableTwoFactorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/me/two-factor/disable": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Turn off 2FA (confirm with password or code; no-op when off)",
        "operationId": "DisableTwoFactor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/me/two-factor/recovery-codes": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Replace all recovery codes (confirm with password or code); returns the new codes once",
        "operationId": "RegenerateRecoveryCodes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "EnableTwoFactorRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "enum": [
          "bad_request",
          "calendar_frozen",
          "conflict",
          "csrf_header_missing",
          "email_domain_not_allowed",
          "email_not_verified",
          "external_sharing_not_allowed",
          "feature_not_in_plan",
          "insufficient_permission",
          "internal_error",
          "invalid_credentials",
          "method_not_allowed",
          "not_found",
          "override_invalid_in_target",
          "payload_too_large",
          "permission_self_lockout",
          "plan_limit_reached",
          "precondition_failed",
          "precondition_required",
          "rate_limited",
          "reauthentication_failed",
          "recurrence_invalid",
          "service_unavailable",
          "time_zone_invalid",
          "token_expired",
          "token_invalid",
          "two_factor_required",
          "uid_conflict",
          "unauthenticated",
          "unsupported_media_type",
          "validation_failed"
        ],
        "type": "string",
        "description": "Stable machine-readable error code (part of the contract; never renamed)."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Validation errors by field (code `validation_failed`)."
          }
        },
        "description": "RFC 9457 problem details. Further members depend on the code (e.g. `limit`, `required`, `actual`)."
      },
      "ReauthenticationRequest": {
        "type": "object",
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "maxLength": 16,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RecoveryCodesResponse": {
        "required": [
          "recoveryCodes"
        ],
        "type": "object",
        "properties": {
          "recoveryCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TwoFactorSetupResponse": {
        "required": [
          "sharedKey",
          "authenticatorUri"
        ],
        "type": "object",
        "properties": {
          "sharedKey": {
            "type": "string"
          },
          "authenticatorUri": {
            "type": "string"
          }
        }
      },
      "TwoFactorStatusResponse": {
        "required": [
          "enabled",
          "recoveryCodesLeft"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "recoveryCodesLeft": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Session cookie set by POST /api/v1/auth/login (web app, same origin). Unsafe methods also need the header `X-Requested-With: scal`.",
        "name": "__Host-scal",
        "in": "cookie"
      }
    }
  },
...
}